fix extra renders 3: desktop tab bar (#22982)
* WIP

* WIP

* WIP

* fix extra renders 4 (#22986)

* WIP

* WIP

* WIP

* WIP
