+++ shared/chat/conversation/messages/attachment/image/index.tsx
@@ -3,6 +3,7 @@ import * as Kb from '../../../../../common-adapters'
 import * as Styles from '../../../../../styles'
 import {ImageRender} from './image-render'
 import {isMobile} from '../../../../../util/container'
+import {memoize} from '../../../../../util/memoize'
 import * as Types from '../../../../../constants/types/chat2'
 import * as Constants from '../../../../../constants/chat2'
 import {ShowToastAfterSaving} from '../shared'
@@ -44,10 +45,10 @@ class ImageAttachment extends React.PureComponent<Props, State> {
   imageRef: any
 
   state = {loaded: false, loadingVideo: 'notloaded', playingVideo: false} as State
-  _setLoaded = () => this.setState({loaded: true})
-  _setVideoLoaded = () => this.setState({loadingVideo: 'loaded'})
+  private setLoaded = () => this.setState({loaded: true})
+  private setVideoLoaded = () => this.setState({loadingVideo: 'loaded'})
 
-  _onClick = () => {
+  private onClick = () => {
     // Once the user clicks the inline video once, then just let the native controls handle everything else.
     if (this.state.playingVideo) {
       return
@@ -62,16 +63,21 @@ class ImageAttachment extends React.PureComponent<Props, State> {
       this.props.onClick()
     }
   }
-  _onDoubleClick = () => {
+  private onDoubleClick = () => {
     if (this.props.inlineVideoPlayable && this.imageRef) {
       this.imageRef.pauseVideo()
     }
     this.props.onDoubleClick()
   }
 
+  private memoizedMeta = memoize((message: Props['message']) => {
+    return {message}
+  })
+
   render() {
     const progressLabel = Constants.messageAttachmentTransferStateToProgressLabel(this.props.transferState)
     const mobileImageFilename = this.props.message.deviceType === 'mobile'
+
     return (
       <>
         <ShowToastAfterSaving transferState={this.props.transferState} />
@@ -107,8 +113,8 @@ class ImageAttachment extends React.PureComponent<Props, State> {
                 {!!this.props.path && (
                   <Kb.Box2 direction="vertical" alignItems="center">
                     <Kb.ClickableBox
-                      onClick={this._onClick}
-                      onDoubleClick={this._onDoubleClick}
+                      onClick={this.onClick}
+                      onDoubleClick={this.onDoubleClick}
                       onLongPress={this.props.toggleMessageMenu}
                     >
                       <Kb.Box2
@@ -127,8 +133,8 @@ class ImageAttachment extends React.PureComponent<Props, State> {
                           }}
                           src={this.props.path}
                           videoSrc={this.props.fullPath}
-                          onLoad={this._setLoaded}
-                          onLoadedVideo={this._setVideoLoaded}
+                          onLoad={this.setLoaded}
+                          onLoadedVideo={this.setVideoLoaded}
                           loaded={this.state.loaded}
                           inlineVideoPlayable={this.props.inlineVideoPlayable}
                           height={this.props.height}
@@ -186,7 +192,7 @@ class ImageAttachment extends React.PureComponent<Props, State> {
                         alignItems="flex-start"
                       >
                         <Kb.Markdown
-                          meta={{message: this.props.message}}
+                          meta={this.memoizedMeta(this.props.message)}
                           selectable={true}
                           allowFontScaling={true}
                         >
