+++ shared/chat/conversation/messages/system-joined/index.tsx
@@ -135,7 +135,7 @@ const AuthorAndAvatar = (props: {
         onUsernameClicked="profile"
         style={Styles.collapseStyles([props.authorIsYou && styles.usernameHighlighted])}
         type="BodySmallBold"
-        usernames={[props.username]}
+        usernames={props.username}
       />
       <Kb.Text type="BodyTiny" style={styles.timestamp}>
         {formatTimeForChat(props.timestamp)}
