+++ shared/chat/conversation/messages/wrapper/exploding-height-retainer/index.desktop.tsx
@@ -108,7 +108,7 @@ const Ashes = (props: {doneExploding: boolean; exploded: boolean; explodedBy?: s
         <Kb.ConnectedUsernames
           type="BodySmallBold"
           onUsernameClicked="profile"
-          usernames={[props.explodedBy]}
+          usernames={props.explodedBy}
           inline={true}
           colorFollowing={true}
           colorYou={true}
