+++ shared/chat/conversation/messages/wrapper/timestamp.tsx
@@ -6,17 +6,13 @@ export type Props = {
   timestamp: string
 }
 
-class Timestamp extends React.PureComponent<Props> {
-  render() {
-    return (
-      <Kb.Box style={styles.box}>
-        <Kb.Text style={styles.text} type="BodySmallSemibold">
-          {this.props.timestamp}
-        </Kb.Text>
-      </Kb.Box>
-    )
-  }
-}
+const Timestamp = React.memo((props: Props) => (
+  <Kb.Box style={styles.box}>
+    <Kb.Text style={styles.text} type="BodySmallSemibold">
+      {props.timestamp}
+    </Kb.Text>
+  </Kb.Box>
+))
 
 const styles = Styles.styleSheetCreate(
   () =>
