+++ shared/chat/inbox/row/small-team/container.tsx
@@ -63,11 +63,14 @@ export default Container.namedConnect(
     const isDecryptingSnippet =
       (hasUnread || stateProps.snippet.length === 0) && Constants.isDecryptingSnippet(stateProps._meta)
     const hasResetUsers = stateProps._meta.resetParticipants.size !== 0
-    const participants = stateProps._participantInfo.all.length
+    const participantsArray = stateProps._participantInfo.all.length
       ? Constants.getRowParticipants(stateProps._participantInfo, stateProps._username)
       : !ownProps.isTeam
       ? ownProps.name.split(',')
       : [ownProps.name]
+
+    const participants = participantsArray.length === 1 ? participantsArray[0] : participantsArray
+
     const teamname = stateProps._meta.teamname
       ? stateProps._meta.teamname
       : ownProps.isTeam
