+++ shared/profile/user/friend/index.tsx
@@ -22,7 +22,7 @@ class Friend extends React.PureComponent<Props> {
           <Kb.Avatar size={64} username={p.username} style={styles.avatar} showFollowingStatus={true} />
           <Kb.ConnectedUsernames
             type={Styles.isMobile ? 'BodySmallBold' : 'BodyBold'}
-            usernames={[p.username]}
+            usernames={p.username}
             colorBroken={true}
             colorFollowing={true}
           />
