+++ shared/router-v2/router.desktop.tsx
@@ -3,7 +3,7 @@ import * as Constants from '../constants/router2'
 import * as Tabs from '../constants/tabs'
 import * as Styles from '../styles'
 import * as React from 'react'
-import TabBar from './tab-bar/container.desktop'
+import TabBar from './tab-bar.desktop'
 import {
   NavigationViewProps,
   createNavigator,
