+++ shared/teams/team/rows/member-row/index.tsx
@@ -98,7 +98,7 @@ export const TeamMemberRow = (props: Props) => {
 
         <Kb.Box2 direction="vertical" fullWidth={true} style={styles.nameContainer}>
           <Kb.Box style={Styles.globalStyles.flexBoxRow}>
-            <Kb.ConnectedUsernames type="BodyBold" usernames={[props.username]} />
+            <Kb.ConnectedUsernames type="BodyBold" usernames={props.username} />
           </Kb.Box>
 
           <Kb.Box style={styles.nameContainerInner}>
@@ -227,7 +227,7 @@ export const TeamMemberRow = (props: Props) => {
           <Kb.Avatar username={props.username} size={Styles.isMobile ? 48 : 32} />
           <Kb.Box style={styles.nameContainer}>
             <Kb.Box style={Styles.globalStyles.flexBoxRow}>
-              <Kb.ConnectedUsernames type="BodyBold" usernames={[props.username]} />
+              <Kb.ConnectedUsernames type="BodyBold" usernames={props.username} />
             </Kb.Box>
             <Kb.Box style={styles.nameContainerInner}>
               {fullNameLabel}
