+++ shared/teams/team/rows/menu-header.new.tsx
@@ -11,7 +11,7 @@ export type Props = {
 const MenuHeader = (props: Props) => (
   <Kb.Box2 direction="vertical" fullWidth={true} alignItems="center" style={styles.header}>
     <Kb.Avatar username={props.username} size={64} style={styles.avatar} />
-    <Kb.ConnectedUsernames type="BodyBold" colorFollowing={true} usernames={[props.username]} />
+    <Kb.ConnectedUsernames type="BodyBold" colorFollowing={true} usernames={props.username} />
     {!!props.fullName && <Kb.Text type="BodySmall">{props.fullName}</Kb.Text>}
     {!!props.label && typeof props.label === 'string' ? (
       <Kb.Text type="BodySmall">{props.label}</Kb.Text>
